Add clean and sync capabilities
Clean flag now allows the csv and pickles for a given
log to be deleted without needing to search for them
yourself. Syncing allows the user to designate a
message to use as a syncing for a shared csv file.
This has file size ramifications. If used with
PlotJuggler, the sync message should be the x axis
ideally. In the future some time interpolation or
or the like may be good.